Add quality selector for progressive download videos
This patch adds the `paella-mp4multiquality-plugin` to bring
back the quality selector for progressive download videos
known from Paella 6. It also update the dependencies to the
versions used in the `engage-paella-player-7` Opencast module.
Lastly, I ran `npm audit fix` to let it fix vulnerabilities.

Changes are analogous to:
* https://github.com/opencast/opencast/pull/5685
* https://github.com/elan-ev/tobira/pull/1143

See also:
* https://github.com/polimediaupv/paella-core/pull/352
* https://github.com/polimediaupv/paella-mp4multiquality-plugin