Add video canvas plugin preventing the context menu on right-click
This patch adds a video plugin allowing to prevent the context menu when
right-clicking within the video container. By default, this plugin is not active.
It has to be configured explicitly within `config.json`. To do so, disable
`es.upv.paella.videoCanvas` in the default configuration and add the following
entry:

```
"org.opencast.paella.videoCanvasWithoutContextMenu": {
  "enabled": true,
  "order": 0
}
```

Fixes #67